refactor(modo-upload): change FileStorage to Arc, extract FromMultipart trait
Switch storage() factory and all handlers from Box<dyn FileStorage> to
Arc<dyn FileStorage> for cheap cloning. Move the FromMultipart trait
definition from lib.rs into its own from_multipart.rs file.
